Added : Use Case 2 Program - Find out the maximum from given three float values

+++ src/com/generics/assignments/TestMaximum.java
@@ -1,4 +1,5 @@
 //Use Case 1 is to give 3 integers and find out the maximum and its position in array.
+//Use Case 2 is to give 3 floats and find out the maximum and its position in array.
 
 package com.generics.assignments;
 
@@ -17,14 +18,14 @@ public class TestMaximum {
         TestMaximum testMaximum = new TestMaximum();
 
         //Define the array where size is 3(given).
-        Integer[] array = addToArray();
+        Float[] array = addToArray();
         printArray(array);
         getMaximum(array);
     }
 
     //Get maximum from the array and its position.
-    public static void getMaximum(Integer[] array) {
-        int max = array[0];
+    public static void getMaximum(Float[] array) {
+        float max = array[0];
         int position = 0;
         for (int i = 0; i < array.length; i++)
         {
@@ -39,8 +40,8 @@ public class TestMaximum {
         System.out.println("Maximum integer position is : "+position+" in array.");
     }
 
-    //Printing to array.
-    public static void printArray(Integer[] array) {
+    //Printing the array.
+    public static void printArray(Float[] array) {
         System.out.println("Array is : ");
         for (int i = 0; i < array.length; i++)
         {
@@ -49,13 +50,13 @@ public class TestMaximum {
     }
 
     //Adding values to array.
-    public static Integer[] addToArray() {
+    public static Float[] addToArray() {
         System.out.println("Enter 3 values");
         Scanner scan = new Scanner(System.in);
-        int a = scan.nextInt();
-        int b = scan.nextInt();
-        int c = scan.nextInt();
+        float a = scan.nextFloat();
+        float b = scan.nextFloat();
+        float c = scan.nextFloat();
         scan.close();
-        return new Integer[]{a,b,c};
+        return new Float[]{a,b,c};
     }
 }
